# swagger.yaml
openapi: 3.0.0
info:
  title: ISDServices
  description: In order to create a post on your medical reserch work, read about other medical post and comments, ISDServices provides you this placform, you must first initiate the process from the platform by registering.
  contact:
    name: hr.admin@isdservices.org
    email: integratedsoftwaredataservices@gmail.com
  version: "1.0"
servers:
  - url: https://isdservices.herokuapp.com
    description: This is the production server
    variables: {}
  - url: http://localhost:4400
    description: "This is the localhost development server"
    variables: {}
paths:
...
  /isds/suscribe:
    post:
      tags:
        - User
      summary: suscribe
      description: suscribe to news letter
      operationId: suscribe
      parameters: []
      requestBody:
        description: "Suscribe to ISDS newsletter"
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/suscribe"
                - example:
                    email: test@domain.com
            example:
              email: test@domain.com
        required: true
      responses:
        "200":
          description: ""
          headers: {}
          content: {}
      deprecated: false
      security: []
...
  /post/create-post:
    post:
      tags:
        - User
      summary: CreatePost
      description: A logged-in user can create a post with this endpoint
      operationId: CreatePost
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/createPost"
                - example:
                    postTitle: Post One
                    postBody: "post body text"
                    postImage: "http://img.png"
                    category: "health care"
                    author: "Bassey Edet"
                    userSummary: "I am full stack web developer"
            example:
              postTitle: Post One
              postBody: "post body text"
              postImage: "http://img.png"
              category: "health care"
              author: "Bassey Edet"
              userSummary: "I am full stack web developer"
        required: true
      responses:
        "200":
          description: ""
          headers: {}
          content: {}
      deprecated: false
      security:
        - bearer: []
...
  /user/all-users:
    get:
      tags:
        - Admin
      summary: AllUsers
      description: This endpoint returns all registered users, accessed by logged-in admin only
      operationId: AllUsers
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
          content: {}
      deprecated: false
      security:
        - bearer: []
  /post/delete-post/{id}:
    delete:
      tags:
        - Admin
      summary: deletePost
      description: This endpoint allows ISDS admin delete a post, accessed by logged-in admin only
      operationId: deletePost
      parameters:
        - name: id
          in: path
          description: The post ID
          required: true
          style: simple
          schema:
            type: string
      responses:
        "200":
          description: "This endpoint allows a varified admin to delete a post"
          headers: {}
          content: {}
      deprecated: false
      security:
        - bearer: []
...
components:
  schemas:
...
    createPost:
      title: createPost
      required:
        - postTitle
        - postBody
        - postImage
        - userSummary
        - userId
        - author
        - category
      type: object
      properties:
        postTitle:
          type: string
        postBody:
          type: string
        postImage:
          type: string
        userSummary:
          type: string
        userId:
          type: string
        category:
          type: string
        author:
          type: string
      example:
        postTitle: Post One
        postBody: "post body text"
        postImage: "https://image.png"
        category: "health care"
        author: "Bassey David"
...
    suscribe:
      title: suscribe
      required:
        - email

      type: object
      properties:
        email:
          type: string

      example:
        email: test@domain.com
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...